Index

Graphons.AbstractGraphon — Type

AbstractGraphon{T,M}

T is the edge type:

Bool for a simple graph

<:Real for a weighted graph

<:AbstractVector{Bool} for a multiplex network

...

M is the type of the representation of the sampled graph from the graphon, e.g. a simple graph with boolean edges can be represented by - a dense matrix of Bool -> M = Matrix{Bool} - a sparse matrix of Bool -> M = SparseMatrixCSC{Bool,Int} - an adjacency list -> M = Vector{Vector{Int}}

for now we assume that M is such that T <: eltype(M) and that it can be created with make_empty_graph(M, n) which creates an empty graph of size n x n

Base.rand — Method

rand(rng::AbstractRNG, f::AbstractGraphon{T,M}, n::Int) where {T,M}

Generates a random graph according to the graphon `f` with `n` nodes.
The latent positions are drawn uniformly at random in [0,1].

The generated graph is of type `M` and has edge type `T`.

Graphons._rand! — Method

_rand!(rng::AbstractRNG, f::AbstractGraphon{T,M}, A::M, ξs)

Generates a random graph according to the graphon `f` and the latent positions `ξs`.
The generated graph is stored in `A`.

Warning

This function expects that A is an empty graph of the right size and type. It does not try to clean it up before filling it. See make_empty_graph for more details.

Graphons.sample — Method

sample(rng::AbstractRNG, f::AbstractGraphon, n::Int)

Generates a random graph according to the graphon `f` with `n` nodes.
The latent positions are drawn uniformly at random in [0,1].

The generated graph is of type `M` and has edge type `T`.